requirements.txt
asgiref==3.5.2
Django==4.1
django-dotenv==1.4.2
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
